requirements.txt
SpeechRecognition>=3.8.1
pyaudio>=0.2.11
numpy>=1.21.0
pymongo>=4.0.0
openai>=0.27.0
pyttsx3>=2.90
scipy>=1.7.0
librosa>=0.9.0
python-dotenv>=0.19.0
sounddevice>=0.4.4
openai-whisper>=1.1.10